+++ fabric-permission-api-v1/src/main/java/net/fabricmc/fabric/mixin/permission/CommandStats_1Mixin.java
@@ -14,17 +14,16 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.mixin.command;
+package net.fabricmc.fabric.mixin.permission;
 
 import org.spongepowered.asm.mixin.Mixin;
 
-import net.minecraft.server.MinecraftServer;
-import net.minecraft.server.network.ServerPlayerEntity;
+import net.fabricmc.fabric.api.permission.v1.PermissibleCommandSource;
 
-@Mixin(ServerPlayerEntity.class)
-public abstract class ServerPlayerEntityMixin extends PlayerEntityMixin {
+@Mixin(targets = "net/minecraft/command/CommandStats$1")
+public abstract class CommandStats_1Mixin implements PermissibleCommandSource {
 	@Override
 	public boolean hasPermission(String perm) {
-		return MinecraftServer.getServer().getPlayerManager().isOperator(this.getGameProfile()) || super.hasPermission(perm);
+		return true;
 	}
 }
