+++ fabric-permission-api-v1/src/main/java/net/fabricmc/fabric/mixin/permission/EntityMixin.java
@@ -14,18 +14,18 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.mixin.command;
+package net.fabricmc.fabric.mixin.permission;
 
 import org.spongepowered.asm.mixin.Mixin;
 
-import net.minecraft.world.CommandBlockExecutor;
+import net.minecraft.entity.Entity;
 
-import net.fabricmc.fabric.api.command.v2.PermissibleCommandSource;
+import net.fabricmc.fabric.api.permission.v1.PermissibleCommandSource;
 
-@Mixin(CommandBlockExecutor.class)
-public abstract class CommandBlockExecutorMixin implements PermissibleCommandSource {
+@Mixin(Entity.class)
+public abstract class EntityMixin implements PermissibleCommandSource {
 	@Override
 	public boolean hasPermission(String perm) {
-		return true;
+		return false;
 	}
 }
