+++ fabric-permission-api-v1/src/main/java/net/fabricmc/fabric/mixin/permission/PlayerEntityMixin.java
@@ -14,27 +14,24 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.mixin.command;
+package net.fabricmc.fabric.mixin.permission;
 
-import org.spongepowered.asm.mixin.Dynamic;
-import org.spongepowered.asm.mixin.Final;
+import com.mojang.authlib.GameProfile;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 
 import net.minecraft.entity.player.PlayerEntity;
 
-import net.fabricmc.fabric.api.command.v2.PermissibleCommandSource;
+import net.fabricmc.fabric.api.permission.v1.PermissibleCommandSource;
 
-@Mixin(targets = "net/minecraft/block/entity/SignBlockEntity$2")
-public abstract class SignBlockEntity_2Mixin implements PermissibleCommandSource {
-	@SuppressWarnings("ShadowTarget")
-	@Final
-	@Dynamic
+@Mixin(PlayerEntity.class)
+public abstract class PlayerEntityMixin implements PermissibleCommandSource {
 	@Shadow
-	PlayerEntity field_1466;
+	public abstract GameProfile getGameProfile();
 
+	// TODO
 	@Override
 	public boolean hasPermission(String perm) {
-		return ((PermissibleCommandSource) this.field_1466).hasPermission(perm);
+		return false;
 	}
 }
