+++ fabric-permission-api-v1/src/main/java/net/fabricmc/fabric/mixin/permission/SignBlockEntity_1Mixin.java
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.mixin.command;
+package net.fabricmc.fabric.mixin.permission;
 
 import org.spongepowered.asm.mixin.Mixin;
 
-import net.fabricmc.fabric.api.command.v2.PermissibleCommandSource;
+import net.fabricmc.fabric.api.permission.v1.PermissibleCommandSource;
 
 @Mixin(targets = "net/minecraft/block/entity/SignBlockEntity$1")
 public abstract class SignBlockEntity_1Mixin implements PermissibleCommandSource {
