+++ fabric-permission-api-v1/src/main/resources/fabric-permission-api-v1.mixins.json
@@ -1,0 +1,20 @@
+{
+  "required": true,
+  "package": "net.fabricmc.fabric.mixin.permission",
+  "compatibilityLevel": "JAVA_8",
+  "mixins": [
+    "CommandBlockExecutorMixin",
+    "CommandStats_1Mixin",
+    "ConsoleMixin",
+    "EntityMixin",
+    "ExecuteCommand_1Mixin",
+    "MinecraftServerMixin",
+    "PlayerEntityMixin",
+    "ServerPlayerEntityMixin",
+    "SignBlockEntity_1Mixin",
+    "SignBlockEntity_2Mixin"
+  ],
+  "injectors": {
+    "defaultRequire": 1
+  }
+}
