+++ fabric-sponge-command-api-v2/src/main/java/net/fabricmc/fabric/api/command/v2/lib/sponge/CommandCallable.java
@@ -34,7 +34,7 @@ import net.minecraft.text.Text;
 import net.minecraft.world.World;
 
 import net.fabricmc.fabric.api.util.Location;
-import net.fabricmc.fabric.api.command.v2.PermissibleCommandSource;
+import net.fabricmc.fabric.api.permission.v1.PermissibleCommandSource;
 import net.fabricmc.fabric.api.command.v2.lib.sponge.spec.CommandSpec;
 
 /**
