+++ fabric-sponge-command-api-v2/src/main/java/net/fabricmc/fabric/impl/command/CommandWrapper.java
@@ -29,7 +29,7 @@ import net.minecraft.text.Text;
 import net.minecraft.util.math.BlockPos;
 
 import net.fabricmc.fabric.api.util.Location;
-import net.fabricmc.fabric.api.command.v2.PermissibleCommandSource;
+import net.fabricmc.fabric.api.permission.v1.PermissibleCommandSource;
 import net.fabricmc.fabric.api.command.v2.lib.sponge.CommandException;
 import net.fabricmc.fabric.api.command.v2.lib.sponge.CommandMapping;
 import net.fabricmc.fabric.api.command.v2.lib.sponge.CommandMessageFormatting;
