+++ fabric-sponge-command-api-v2/src/main/java/net/fabricmc/fabric/impl/command/lib/sponge/args/PlayerCommandElement.java
@@ -36,7 +36,7 @@ import net.minecraft.server.MinecraftServer;
 import net.minecraft.text.LiteralText;
 import net.minecraft.text.Text;
 
-import net.fabricmc.fabric.api.command.v2.PermissibleCommandSource;
+import net.fabricmc.fabric.api.permission.v1.PermissibleCommandSource;
 import net.fabricmc.fabric.api.command.v2.lib.sponge.args.ArgumentParseException;
 import net.fabricmc.fabric.api.command.v2.lib.sponge.args.CommandArgs;
 import net.fabricmc.fabric.api.command.v2.lib.sponge.args.SelectorCommandElement;
